+++ features/step_definitions/web_steps/i18n_steps.rb
@@ -1,5 +1,5 @@
 Given /^i18n key "(.*?)" with value "(.*?)"$/ do |keys, value|
-  data = {}.store_multi(value, *keys.split('.'))
+  data = {}.dig_store(value, *keys.split('.'))
   I18n.backend.store_translations(I18n.locale, data)
 end